Module Plist_intf.Pl_origin

This type is the standard information we need to persist in order to quickly restore the plist; we can also just follow the list from the hd block of course, but this is O(n).

type 'blk_id pl_origin = {

hd : 'blk_id;

tl : 'blk_id;

blk_len : int;

}

val bin_shape_pl_origin : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_pl_origin : blk_id. 'blk_id Bin_prot.Size.sizer -> 'blk_id pl_origin Bin_prot.Size.sizer

val bin_write_pl_origin : blk_id. 'blk_id Bin_prot.Write.writer -> 'blk_id pl_origin Bin_prot.Write.writer

val bin_writer_pl_origin : 'a Bin_prot.Type_class.writer -> 'a pl_origin Bin_prot.Type_class.writer

val __bin_read_pl_origin__ : blk_id. 'blk_id Bin_prot.Read.reader -> (int -> 'blk_id pl_origin) Bin_prot.Read.reader

val bin_read_pl_origin : blk_id. 'blk_id Bin_prot.Read.reader -> 'blk_id pl_origin Bin_prot.Read.reader

val bin_reader_pl_origin : 'a Bin_prot.Type_class.reader -> 'a pl_origin Bin_prot.Type_class.reader

val bin_pl_origin : 'a Bin_prot.Type_class.t -> 'a pl_origin Bin_prot.Type_class.t

val mshlr : r_mshlr:'r Tjr_fs_shared.bp_mshlr -> 'r pl_origin Tjr_fs_shared.bp_mshlr

type 'blk_id t = 'blk_id pl_origin

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : blk_id. 'blk_id Bin_prot.Size.sizer -> 'blk_id t Bin_prot.Size.sizer

val bin_write_t : blk_id. 'blk_id Bin_prot.Write.writer -> 'blk_id t Bin_prot.Write.writer

val bin_writer_t : 'a Bin_prot.Type_class.writer -> 'a t Bin_prot.Type_class.writer

val __bin_read_t__ : blk_id. 'blk_id Bin_prot.Read.reader -> (int -> 'blk_id t) Bin_prot.Read.reader

val bin_read_t : blk_id. 'blk_id Bin_prot.Read.reader -> 'blk_id t Bin_prot.Read.reader

val bin_reader_t : 'a Bin_prot.Type_class.reader -> 'a t Bin_prot.Type_class.reader

val bin_t : 'a Bin_prot.Type_class.t -> 'a t Bin_prot.Type_class.t

type ('blk_id, 't) ops = {

read : unit -> 'blk_id t;

write : 'blk_id t -> (unit, 't) Tjr_monad.m;

set_and_sync : 'blk_id t -> (unit, 't) Tjr_monad.m;
convenience combination of set and sync

}