Module Plist_intf.Plist_ops

type ('a, 'buf, 'blk_id, 't) plist_ops = {

add : nxt:'blk_id -> elt:'a -> ('blk_id option, 't) Tjr_monad.m;

add_if_room : 'a -> (bool, 't) Tjr_monad.m;

sync_tl : unit -> (unit, 't) Tjr_monad.m;

adv_hd : unit -> (('a, 'blk_id) adv_hd or_error, 't) Tjr_monad.m;

adv_tl : 'blk_id -> (unit, 't) Tjr_monad.m;

blk_len : unit -> (int, 't) Tjr_monad.m;

get_origin : unit -> ('blk_id Pl_origin.pl_origin, 't) Tjr_monad.m;

read_hd : unit -> ('a list * 'blk_id option, 't) Tjr_monad.m;

append : ('blk_id, 'buf) plist -> (unit, 't) Tjr_monad.m;

}

plist operations which require the plist state from the monad

add : The blk_id is returned if it is not used; another variant assumes an alloc function. We don't assume the nxt_blk is clean - we may write a single elt into the new blk then barrier, before updating the nxt pointer.

add_if_room: returns a boolean to indicate if the element was added in the current tl

sync : we automatically barrier before moving to a new tl (FIXME check); this is for partial blocks which we need to sync (eg for clean shutdown)

adv_hd: return an error if no nxt pointer (iff blk_len = 1 iff hd=tl)

adv_tl: write new empty tl, update old tl, and advance tl to new

read_hd: assumes hd has been marshalled to disk; this likely won't be the case if hd=tl FIXME perhaps keep hd in mem? why do we need this?

append: take a second plist, and adjust the "current" plist so that the hd is unchanged, but the tl points to the tl of the second plist (and the buf and offset etc are copied); an assumption is that the second plist is never used from this point onwards. FIXME clearly we need to remove the pl2 from the global state, or else we have a space leak