Module Make_v1.B

Incorporate block device

Parameters

S2 : sig ... end

Signature

val (>>=) : ('a, S.t) Tjr_monad.m -> ('a -> ('b, S.t) Tjr_monad.m) -> ('b, S.t) Tjr_monad.m

val return : 'a -> ('a, S.t) Tjr_monad.m

val write0 : blk_id:S.blk_id -> blk:S.blk -> (unit, S.t) Tjr_monad.m

val read : blk_id:S.blk_id -> (S.blk, S.t) Tjr_monad.m

val write_and_barrier : blk_id:S.blk_id -> blk:S.blk -> (unit, S.t) Tjr_monad.m

val mk_empty : S.blk_id -> ((S.blk_id, S.buf) Tjr_plist.Plist_intf.plist, S.t) Tjr_monad.m

Make an empty list at blk_id

val read_from_head : S.blk_id -> ((S.a list * S.blk_id option) list, S.t) Tjr_monad.m

Read the plist from disk, starting at the head; O(n) in length of list

val init_from_endpts : hd:S.blk_id -> tl:S.blk_id -> blk_len:int -> ((S.blk_id, S.buf) Tjr_plist.Plist_intf.plist, S.t) Tjr_monad.m

Restore the plist given hd and tl pointers (where tl may actually have nxt set, since tl is usually just the last SYNCED tl); O(delta) where delta is the length of the list after the tl block.

val init : < create : S.blk_id -> ((S.blk_id, S.buf) Tjr_plist.Plist_intf.plist, S.t) Tjr_monad.m; from_endpts : S.blk_id Tjr_plist.Plist_intf.Pl_origin.pl_origin -> ((S.blk_id, S.buf) Tjr_plist.Plist_intf.plist, S.t) Tjr_monad.m; read_from_hd : S.blk_id -> ((S.a list * S.blk_id option) list, S.t) Tjr_monad.m; >

module C : functor (S3 : sig ... end) -> sig ... end

Incorporate state passing