Module Plist_intf.Pl_origin

This type is the standard information we need to persist in order to quickly restore the plist; we can also just follow the list from the hd block of course, but this is O(n).

type 'blk_id pl_origin = {

hd : 'blk_id;

tl : 'blk_id;

blk_len : int;

}

type 'blk_id t = 'blk_id pl_origin

type ('blk_id, 't) ops = {

set_and_sync : 'blk_id t -> (unit, 't) Tjr_monad.m;
convenience combination of set and sync

}