Module Tjr_plist__Make_2

A persistent list for arbitrary elts, assuming existence of a freelist (so we don't have to closely track blk_ids)

type ('a, 'blk_id, 't) simple_plist_ops = {

add : 'a -> (bool, 't) Tjr_monad.m;
Return value indicates whether we moved to a new block

sync_tl : unit -> (unit, 't) Tjr_monad.m;

blk_len : unit -> (int, 't) Tjr_monad.m;

get_origin : unit -> ('blk_id Tjr_plist.Plist_intf.Pl_origin.pl_origin, 't) Tjr_monad.m;

}

type ('a, 'blk_id, 'blk, 'buf, 't) simple_plist_factory = < plist_factory : ('a, 'blk_id, 'blk, 'buf, 't) Tjr_plist.Plist_intf.plist_factory; convert_to_simple_plist : monad_ops:'t Tjr_monad.monad_ops -> freelist_ops:('blk_id, 't) Tjr_fs_shared.Shared_freelist.freelist_ops -> plist_ops:('a, 'buf, 'blk_id, 't) Tjr_plist.Plist_intf.plist_ops -> ('a, 'blk_id, 't) simple_plist_ops; >

Plist_ops_2 from plist_ops

val convert_to_simple_plist : monad_ops:'a Tjr_monad.monad_ops -> freelist_ops:('b, 'a) Tjr_fs_shared.Shared_freelist.freelist_ops -> plist_ops:('c, 'd, 'b, 'a) Tjr_plist.Plist_intf.plist_ops -> ('c, 'b, 'a) simple_plist_ops

val examples : < for_int : (int, Tjr_plist__.Pl_type_abbrevs.blk_id, Tjr_plist__.Pl_type_abbrevs.blk, Tjr_fs_shared.ba_buf, Tjr_plist__.Pl_type_abbrevs.t) simple_plist_factory; for_int_int_kvop : ((int, int) Tjr_fs_shared.kvop, Tjr_plist__.Pl_type_abbrevs.blk_id, Tjr_plist__.Pl_type_abbrevs.blk, Tjr_fs_shared.ba_buf, Tjr_plist__.Pl_type_abbrevs.t) simple_plist_factory; >