Module Tjr_plist__Plist_intf.Pl_root_info

type 'blk_id pl_root_info = {

hd : 'blk_id;

tl : 'blk_id;

blk_len : int;

}

This type is the standard information we need to persist in order to quickly restore the plist; we can also just follow the list from the hd block of course, but this is O(n).