Module Tjr_plist__Plist_sync_root_blk

Utility module, for adding sync when hd,tl ptrs change

module Plist_sync_root_blk : sig ... end

type ('blk_id, 't) srb = ('blk_id, 't) Plist_sync_root_blk.t

val plist_add_sync : monad_ops:'a Tjr_monad.monad_ops -> sync_root_blk:(('b * 'b) -> (unit, 'a) Tjr_monad.m) -> plist_ops:('c, 'd, 'b, 'a) Tjr_plist.Plist_intf.plist_ops -> ('c, 'd, 'b, 'a) Tjr_plist.Plist_intf.plist_ops

Utility: this adds sync operations whenever hd or tl change; not thread safe - assumes only one thread is accessing the list (but probably works even in the multithreaded case, except that the root syncs may be out of order)